add role validation

+++ src/main/java/br/insper/conexoes/connections/ConnectionController.java
@@ -1,7 +1,10 @@
 package br.insper.conexoes.connections;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -38,9 +41,16 @@ public class ConnectionController {
 
     @DeleteMapping
     public void deleteConnection(
-            @RequestHeader(name = "Authorization") String token,
+            @AuthenticationPrincipal Jwt jwt,
             @RequestBody DeleteConnectionRequest request) {
-        System.out.println(token);
+        List<String> roles = jwt.getClaimAsStringList("https://social-insper.com/roles");
+
+        if (roles == null || !roles.contains("ADMIN")) {
+            throw new ResponseStatusException(
+                    HttpStatus.FORBIDDEN,
+                    "Apenas ADMIN pode executar esta ação"
+            );
+        }
         service.delete(
                 request.fromUserId(),
                 request.toUserId()
